+++ plugins/data/dataframes/src/hopeit/dataframes/serialization/dataset.py
@@ -1,11 +1,15 @@
 """Dataset objects definition, used as a result of serialized dataframes"""
 
 from importlib import import_module
-from typing import Generic, Type, TypeVar
+from typing import Any, Dict, Generic, Type, TypeVar
 
-from hopeit.dataobjects import dataclass, dataobject
+from hopeit.dataobjects import dataclass, dataobject, field
 
 DataFrameT = TypeVar("DataFrameT")
+
+
+class DatasetLoadError(Exception):
+    pass
 
 
 @dataobject
@@ -17,9 +21,17 @@ class Dataset(Generic[DataFrameT]):
     partition_key: str
     key: str
     datatype: str
+    schema: Dict[str, Any] = field(default_factory=dict)
 
     async def load(self) -> DataFrameT:
-        return await self.__storage.load(self)  # type: ignore[attr-defined]
+        try:
+            dataframe = await self.__storage.load(self)  # type: ignore[attr-defined]
+            return dataframe
+        except (RuntimeError, IOError, KeyError) as e:
+            raise DatasetLoadError(
+                f"Error {type(e).__name__}: {e} loading dataset of type {self.datatype} "
+                f"at location {self.partition_key}/{self.key}"
+            ) from e
 
     @classmethod
     async def save(cls, dataframe: DataFrameT) -> "Dataset[DataFrameT]":