+++ docs/source/release-notes.rst
@@ -1,6 +1,23 @@
 Release Notes
 =============
 
+Version 0.25.1
+______________
+
+- Engine
+
+  - API: Support for `type | None` syntax, as well as `Optional[type]` annotation for optional required query args
+
+- Plugins:
+
+  - DataFrames, feature to allow schema evolution:
+      - Allowing optional and default values in `@dataframe` objects
+      - Saving `json` schema for serialized Datasets
+      - Handling load errors with `DatasetLoadException`
+      - Allowing load dataset with deprecated fields (removed fields)
+      - Allowing load dataset with new fields in schema using a default value
+
+
 Version 0.25.0
 ______________